Refactor analysisd conftests
